% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_log_moi_proposal_ratio.R
\name{calculate_log_moi_proposal_ratio}
\alias{calculate_log_moi_proposal_ratio}
\title{Calculate log MOI proposal ratio}
\usage{
calculate_log_moi_proposal_ratio(moi_max, moi_current, moi_proposed,
  moi_current_equal_moi_min, moi_proposed_equal_moi_min, datasampleID, no_total)
}
\description{
Function to calculate the log MOI proposal ratio
}

