% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_genotype_count_compatibility.R
\name{data_genotype_count_compatibility}
\alias{data_genotype_count_compatibility}
\title{Data genotype count compatibility}
\usage{
data_genotype_count_compatibility(comp_genotypes, genotype_counts, raw_data)
}
\description{
Data and genotype compatibility check function
}
