% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddirichlet.R
\name{ddirichlet}
\alias{ddirichlet}
\title{Dirichlet density}
\usage{
ddirichlet(x, alpha)
}
\description{
Calculate the density of the dirichlet distribution. This function is exactly the same as that from MCMCpack, copied to minimise dependencies.
}
