% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_frequency_proposed_child.R
\name{draw_frequency_proposed_child}
\alias{draw_frequency_proposed_child}
\title{Draw frequency proposed child}
\usage{
draw_frequency_proposed_child(Alpha, genotype_counts_current)
}
\description{
Function for drawing frequency proposed per child
}
