% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_genotype_counts_initial.R
\name{draw_genotype_counts_initial}
\alias{draw_genotype_counts_initial}
\title{Draw initial genotype counts}
\usage{
draw_genotype_counts_initial(
  moi_current,
  processed_data_list,
  frequency_current
)
}
\description{
Function for drawing initial genotype counts.
}
