% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_NGS.R
\name{log_likelihood_NGS}
\alias{log_likelihood_NGS}
\title{Log likelihood, NGS data}
\usage{
log_likelihood_NGS(processed_data_list, genotype_counts)
}
\description{
Calculate log likelihood of NGS data
}
