% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_prevalence.R
\name{log_likelihood_prevalence}
\alias{log_likelihood_prevalence}
\title{Log likelihood, prevalence data}
\usage{
log_likelihood_prevalence(processed_data_list, genotype_counts)
}
\description{
Calculate log likelihood of prevalence data
}
