% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_moi_prior_calculation_pois.R
\name{log_moi_prior_calculation_pois}
\alias{log_moi_prior_calculation_pois}
\title{Log moi prior}
\usage{
log_moi_prior_calculation_pois(moi_store, moi_list, y_no_mxed, y_mxed)
}
\description{
Log moi prior calculation
}
