% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prior_moi_ratio_geom.R
\name{log_prior_moi_ratio_geom}
\alias{log_prior_moi_ratio_geom}
\title{lLg prior MOI ratio}
\usage{
log_prior_moi_ratio_geom(moi_proposed, moi_current, moi_hyperparameter,
  moi_size_hyperparameter)
}
\description{
Function to calculate the log of the prior MOI ratio when the MOI prior distribution is geometric.
Note that the normalising constants due to truncation cancel.
}

