% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prior_moi_ratio_unif.R
\name{log_prior_moi_ratio_unif}
\alias{log_prior_moi_ratio_unif}
\title{Log prior MOI ratio}
\usage{
log_prior_moi_ratio_unif(
  moi_proposed,
  moi_current,
  moi_hyperparameter,
  moi_size_hyperparameter
)
}
\description{
Function to calculate the log of the prior MOI ratio when the MOI prior distribution is a discrete uniform distribution.
Note that the normalising constants due to truncation cancel.
}
