% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampling.R
\name{mcmc_sampling}
\alias{mcmc_sampling}
\title{MCMC}
\usage{
mcmc_sampling(processed_data_list, moi_list, frequency_list, mcmc_variable_list)
}
\description{
Function to implement mcmc sampler, which is embedded within data preprocessing and retrospective posterior calculations
}

