% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_log_dmultinom.R
\name{my_log_dmultinom}
\alias{my_log_dmultinom}
\title{Log Multinomial Density}
\usage{
my_log_dmultinom(x, prob, pop_fequency)
}
\description{
Function to calculate the density of multiple realisations of the multinomial distribution
simultaneously given either a single probability or a matrix of probability
}
