% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdirichlet.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Random draw from the dirichlet}
\usage{
rdirichlet(n, alpha)
}
\description{
Same as the rdirichlet function in MCMCpack but rewritten to minimise dependencies.
}
\details{
Need to add log option
}

