% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_genotype_counts.R
\name{update_genotype_counts}
\alias{update_genotype_counts}
\title{Update genotype counts}
\usage{
update_genotype_counts(moi_diff, genotype_counts_current,
  current_genotype_counts_truncated, processed_data_list)
}
\description{
Function to update genotype counts conditional on current genotype counts and new multiplicity of infection (MOI)
}

