% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_y_mixed_add.R
\name{update_y_mixed_add}
\alias{update_y_mixed_add}
\title{Update mixed data samaple clone addition}
\usage{
update_y_mixed_add(to_update, genotype_counts_mixed_add, frequency_truncated)
}
\description{
Function for addition of clone to a mixed data sample
}

