% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_y_mixed_sub.R
\name{update_y_mixed_sub}
\alias{update_y_mixed_sub}
\title{Update mixed data samaple clone removal}
\usage{
update_y_mixed_sub(to_update, genotype_counts_mixed_sub,
  genotype_counts_truncated)
}
\description{
Function for removal of clone from mixed data sample
}

