% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatible_genotypes.R
\name{compatible_genotypes}
\alias{compatible_genotypes}
\title{Compatible genotypes}
\usage{
compatible_genotypes(observation, as_numeric = TRUE)
}
\description{
Function to ascertain the genotypes that are compatible with a given observation
}

