% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampling_parallel.R
\name{mcmc_sampling_parallel}
\alias{mcmc_sampling_parallel}
\title{MCMC}
\usage{
mcmc_sampling_parallel(processed_data_list, moi_list, frequency_list,
  mcmc_variable_list, cores_max)
}
\description{
Function to implement mcmc sampler chains in parallel, followed retrospective posterior calculations
}

